refactor: 용어 도메인 DTO 패키지 및 네이밍 변경
