+++ space-d/src/main/java/com/dnd/spaced/core/word/domain/dto/PopularWord.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.word.domain.repository.dto;
+package com.dnd.spaced.core.word.domain.dto;
 
 public record PopularWord(int rank, Long wordId, String name) {
 }
