+++ space-d/src/main/java/com/dnd/spaced/core/word/domain/dto/ViewCountStatisticsRank.java
@@ -1,0 +1,4 @@
+package com.dnd.spaced.core.word.domain.dto;
+
+public record ViewCountStatisticsRank(int rank, Long wordId, long viewCount) {
+}
