--- space-d/src/main/java/com/dnd/spaced/core/word/domain/dto/ViewCountStatisticsRankDto.java
@@ -1,4 +1,0 @@
-package com.dnd.spaced.core.word.domain.dto;
-
-public record ViewCountStatisticsRankDto(int rank, Long wordId, long viewCount) {
-}
